+++ src/interfaces/components/index.ts
@@ -74,7 +74,6 @@
  * Here is a description of the category TextArea.
  *
  * @categoryDescription Toast
- *
  * Here is a description of the category Toast.
  *
  * @categoryDescription ToggleButton
@@ -84,6 +83,10 @@
  * @categoryDescription IconButton
  *
  * Here is a description of the category IconButton.
+ *
+ * @categoryDescription Link
+ *
+ * Here is a description of the category Link.
  *
  * @module Components
  */
@@ -112,6 +115,7 @@ import {
   TextAreaGroup,
   ToggleButtonGroup,
   IconButtonGroup,
+  LinkGroup,
 } from "./Component";
 import { IInput, InputSize, InputType, InputAutocomplete } from "./IInput";
 import { ICheckbox } from "./ICheckbox";
@@ -120,6 +124,7 @@ import { IToast, ToastType } from "./IToast";
 import { IButton, ButtonSize } from "./IButton";
 import { ISkeleton } from "./ISkeleton";
 import { IIconButton } from "./IIconButton";
+import { ILink, LinkType, LinkTarget } from "./ILink";
 import {
   TSelector,
   TBaseSelector,
@@ -171,6 +176,10 @@ export {
   ICreateDialog,
   ISkeleton,
   IIconButton,
+  ILink,
+  LinkType,
+  LinkTarget,
+  LinkGroup,
   TSelector,
   TBaseSelector,
   TFilesSelector,