+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Add actions navigate, openInfoPanel
 - Add IArticleButtonPlugin, IArticleButtonItem
 - Add IconButton component
+- Add ILink component
 - Add IMediaViewer component, actions (showMediaViewer, updateMediaViewer,
   closeMediaViewer)